Makefile: re-add removed SHELL_SCRIPTS - instead ignore missing files when running syntax check; disable examples/more-tests/tests/* and examples/simple/tests/FAIBASE_TEST
Instead of removing the files from SHELL_SCRIPTS that are
available in 3.4-releases but not in trunk versions anymore let's
just check for the existence of the files and if they can't be read
let's skip them. This allows us to check all existing files
independent of the FAI version. This reverts change of svn r6262.

Also drop tests for examples/more-tests/tests/* and
examples/simple/tests/FAIBASE_TEST as they are failing due to:

  Can't locate Faitest.pm in @INC

for now.