add initramfs-tools, workaround for fai-mirror problem
When creating a mirror using fai-mirror, the package linux-image-amd64
will not be added to the mirror because of aptitude's dependency
resolver. This worked with aptitude from lenny, but the squeeze
version will refuse to install/download the kernel package.
Using apt-get will also work, but currently adding initramfs-tools
will fix this issue.