Always install new maintainer config
Changes the default configfile handling of dpkg during update and
install procidures from "--force-confold" to "--force-confnew".

The reasoning behind this change is the problem to restore the default
configuration, if it was ever modified, e.g. by a previous FAI run.

Now the default configuration is restored, if a file was removed from
config space.

Signed-off-by: Thomas Lange <lange@informatik.uni-koeln.de>