fix for upstart support, disks are not detected properly
In the case FAI is started in an Upstart environment, check if an
upstart job udevtrigger exists. If it does, we really need to start
this job so that "missed" kernel events are replayed. Otherwise may
miss block events. This results in block devices having "wrong" owners
and confuses the FAI "disk-info" utility, which in turn breaks the
disk detection completely. LP: #941959

Author: Reinhard Tartler <siretart@tauware.de>