move update of %FAI::disk_var from create_fstab_line() to generate_fstab()
Updating $FAI::disk_var{ROOT_PARTITION} and $FAI::disk_var{SWAPLIST} in
create_fstab_line() is a bit surprising, because the function name does
not indicate such a thing may happen. The function generate_fstab() already
updates $FAI::disk_var{BOOT_DEVICE} so it's better to consolidate the
modifications to %FAI::disk_var into one place.

This is the patch is the first one in a series to replace the global
variable with a suitable return value.