convert find_boot_mnt_point() to use argument instead of %FAI::configs
old function signature:

  my $boot_mnt_point = find_boot_mnt_point();

new function signature:

  my $boot_mnt_point = find_boot_mnt_point(%FAI::config);

This change does not change the actual functionality. Providing the
configuration via a function argument makes it possible to get rid
of the global variable %FAI::configs later on. As an immediate
benefit it is now obvious that find_boot_mnt_point() uses the data
stored in %FAI::configs to find the boot device.