replace [ "$verbose" ] with [ X$verbose = X1 ]
currently the variable $verbose is unset or set to 1. That's why the
old test works. If we want to set verbose to the default value 0,
this test will fail, since it will become true. The new version of
the test will only be true if $verbose is set to 1, and it does not
fail if $verbose is unset.