fai-cd: fix empty expansion causing superflous 1 exit
The last thing fai-cd does in some cases is to move files expanded with
a glob. When creating an autodiscover iso, the program calls itself, and
when it returns, it tries the glob expansion again, but it expands to
nothing, causing the mv to fail. Instead, check that at least one file
exists before doing the move.