fix wrong Perl dereferencing, Closes: #925469
repartitionning using a lvm configuration file and the option preserve_reinstall fails with the message : "Command had non-zero exit code"

The log also says:
Previous partitions overflow begin of preserved partition /dev/sda2

It seems that $FAI::configs does not contain a reference to a
dictionary but is a dictionary object, so dereferencing it leads to an
empty value. As a consequence VG is never preserved by the
preserve_reinstall option. And the last command during the
partitionning is a "vgchange -a n VG", that leads to inactivate
VG. Then when the command blkid is called to generate the /etc/fstab,
it uses a VG that is not activated and the blkid command return an
error code 2. Thus, setup-storage fails.