remove wait_for_jobs() and jobrunning()
jobsrunning did not detect the processes correctly when using
fai-diskimage. It may also find other running processes of the host
that do not belong to the installation.
The only package I know that runs a postinst in the background is
auctex. You can change this behaviour with a preseed:
auctex  auctex/doauto   select  Foreground