Allow . characters in the name of base images
When considering the candidate base images at $FAI_BASEFILEURL, allow
the image filenames to contain a '.' character that is not necessarily a
delimiter for the file extension.  With this change, one could, for
example, set

FAI_BASEFILEURL=https://cloud-images.ubuntu.com/releases/server/focal/release/

and define "ubuntu-20.04-server-cloudimg-amd64-root" as one of the
classes, which would cause FAI to unpack an official Ubuntu cloud image
as the base for an installation.