do not call mkfs multiple times for btrfs with multiple subvolumes in one partition
There is another bug in setup-storage that strikes with almost the same
configuration as above, but when using for example raid1 instead of
single. In that case, mkfs.btrfs is called several times on the same
raid devices, so that previously created subvolumes get lost. This is
because, in contrast to single device btrfs filesystems, setup-storage
does not check for raid if the highest level filesystem already exists.