Save any updated time to hardware
If a machine boots fai, with a hardware clock that's wrong,
it can be set correctly by either rdate or ntpdate.

The issue is that with a modern Debian 12, the later call to udevadm
trigger will trigger a:
/usr/lib/udev/hwclock-set /dev/rtc0

Where the wrong time is re-read from hardware, thus overwriting the
newly corrected time.

This prevent this by just saving the time to hardware after any
rdate/ntpdate calls.