Special handling if a negation is detected
When in download-only mode and a negation is used in the boolean
expression of the PACKAGES line, we have to try to download the
package list, even it may contain package names that are not known for
the current sources.list. fai-mirror may call install_packages with a
long list of classes, so not every logical expression becomes true.